+++ dessia_common/__init__.py
@@ -1,9 +1,6 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
-"""
-__init__ for dessia_common module
-
-"""
+""" __init__ for dessia_common module. """
 
 import pkg_resources