+++ dessia_common/core.py
@@ -1,8 +1,6 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
-"""
-Module to handle serialization for engineering objects.
-"""
+""" Module to handle serialization for engineering objects. """
 
 import time
 import warnings
@@ -25,7 +23,7 @@ from importlib import import_module
 from ast import literal_eval
 
 import dessia_common.errors
-from dessia_common.utils.diff import data_eq, diff, dict_hash, list_hash
+from dessia_common.utils.diff import data_eq, diff, choose_hash
 from dessia_common.utils.types import is_sequence, is_bson_valid, TYPES_FROM_STRING
 from dessia_common.utils.copy import deepcopy_value
 from dessia_common.utils.jsonschema import default_dict, jsonschema_from_annotation, JSONSCHEMA_HEADER, \
@@ -55,7 +53,7 @@ class DessiaObject(SerializableObject):
     Gathers generic methods and attributes
 
     :cvar bool _standalone_in_db:
-        Indicates wether class objects should be independant in database or not.
+        Indicates wether class objects should be independent in database or not.
         If False, object will only exist inside its parent.
 
     :cvar bool _eq_is_data_eq:
@@ -112,6 +110,12 @@ class DessiaObject(SerializableObject):
 
     def __init__(self, name: str = '', **kwargs):
         self.name = name
+        if kwargs:
+            warnings.warn(('Providing attributes to DessiaObject __init__ to be stored in self is deprecated\n'
+                           + 'Please store your attributes by yourself in your init'),
+                          DeprecationWarning)
+
+        # The code below has shown to be unefficient and should be remove in future version (0.16?)
         for property_name, property_value in kwargs.items():
             setattr(self, property_name, property_value)
 
@@ -153,18 +157,8 @@ class DessiaObject(SerializableObject):
 
     def _data_hash(self):
         """ Generic computation of hash based on data. """
-        hash_ = 0
         forbidden_keys = (self._non_data_eq_attributes + self._non_data_hash_attributes + ['package_version', 'name'])
-        for key, value in self._serializable_dict().items():
-            if key not in forbidden_keys:
-                if is_sequence(value):
-                    hash_ += list_hash(value)
-                elif isinstance(value, dict):
-                    hash_ += dict_hash(value)
-                elif isinstance(value, str):
-                    hash_ += sum(ord(v) for v in value)
-                else:
-                    hash_ += hash(value)
+        hash_ = sum(choose_hash(v) for k, v in self._serializable_dict().items() if k not in forbidden_keys)
         return int(hash_ % 1e5)
 
     def _data_diff(self, other_object):
@@ -386,7 +380,7 @@ class DessiaObject(SerializableObject):
         check_list = CheckList([])
 
         if check_platform:
-            check_list += self._check_platform(level=level)
+            check_list += self.check_platform(level=level)
 
         # Type checking: not ready yet
         # class_argspec = inspect.getfullargspec(self.__class__)
@@ -403,7 +397,7 @@ class DessiaObject(SerializableObject):
         return check_list
 
     def is_valid(self, level='error'):
-        """ Return whether the object of valid 'above' given level. Default is error, but warnings can be forbidden. """
+        """ Return whether the object of valid 'above' given level. Default is error, but warning can be forbidden. """
         return not self.check_list().checks_above_level(level=level)
 
     def copy(self, deep: bool = True, memo=None):
@@ -559,6 +553,9 @@ class DessiaObject(SerializableObject):
         print('\n')
 
     def _check_platform(self, level='error'):
+        return self.check_platform().raise_if_above_level(level=level)
+
+    def check_platform(self, level='error'):
         """ Reproduce lifecycle on platform (serialization, display). Raise an error if something is wrong. """
         checks = []
         try:
@@ -566,30 +563,28 @@ class DessiaObject(SerializableObject):
         except TypeError:
             dict_ = self.to_dict()
         json_dict = json.dumps(dict_)
+
         decoded_json = json.loads(json_dict)
         deserialized_object = self.dict_to_object(decoded_json)
+
         if not deserialized_object._data_eq(self):
             print('data diff: ', self._data_diff(deserialized_object))
-            # raise dessia_common.errors.DeserializationError('Object is not equal to itself'
-            #                                                 ' after serialization/deserialization')
             checks.append(FailedCheck('Object is not equal to itself after serialization/deserialization'))
+
         copied_object = self.copy()
         if not copied_object._data_eq(self):
             try:
                 print('data diff: ', self._data_diff(copied_object))
             except:
                 pass
-            checks.append(FailedCheck('Object is not equal to itself after serialization/deserialization'))
-            # raise dessia_common.errors.CopyError('Object is not equal to itself after copy')
+            checks.append(FailedCheck('Object is not equal to itself after copy'))
 
         valid, hint = is_bson_valid(stringify_dict_keys(dict_))
         if not valid:
-            # raise ValueError(hint)
             checks.append(FailedCheck(f'Object is not bson valid {hint}'))
 
         json.dumps(self._displays())
         json.dumps(self._method_jsonschemas)
-
         return CheckList(checks)
 
     def to_xlsx(self, filepath: str):