+++ dessia_common/errors.py
@@ -1,8 +1,6 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
-"""
-Module for errors in dessia_common.
-"""
+""" Module for errors in dessia_common. """
 
 
 class ExceptionWithTraceback(Exception):