+++ dessia_common/utils/copy.py
@@ -1,6 +1,4 @@
-"""
-Tools for copying objects.
-"""
+""" Tools for copying objects. """
 
 import copy
 import warnings
@@ -79,7 +77,7 @@ def deepcopy_value(value, memo):
 
 
 def deepcopy_dict(dict_value, memo):
-    """ Deepcopies a dict. """
+    """ Deepcopy dict. """
     memo_value = search_memo(dict_value, memo)
     if memo_value is not None:
         return memo_value
@@ -93,7 +91,7 @@ def deepcopy_dict(dict_value, memo):
 
 
 def deepcopy_sequence(seq_value, memo):
-    """ Deepcopies a sequence. """
+    """ Deepcopy sequence. """
     memo_value = search_memo(seq_value, memo)
     if memo_value is not None:
         return memo_value
@@ -107,7 +105,7 @@ def deepcopy_sequence(seq_value, memo):
 
 def search_memo(value, memo):
     """ Search in given memo. """
-    for key in memo.keys():
-        if isinstance(value, type(key)) and value == key:
-            return memo[value]
-    return None
+    try:
+        return memo[value]
+    except (TypeError, KeyError):
+        return None